requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
google-genai
pydantic_settings
firebase_admin